requirements.txt
aniso8601==8.0.0
bcrypt==4.0.1
cachelib==0.10.2
certifi==2019.9.11
chardet==3.0.4
click==8.1.3
Flask==2.2.3
Flask-MySQL==1.5.2
Flask-Session==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.21.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.10
text-unidecode==1.3
urllib3==1.25.6
Werkzeug==2.2.3
